+++ feature-statistics/src/main/java/com/github/belyakovleonid/feature_statistics/domain/model/StatisticsCategory.kt
@@ -3,7 +3,7 @@ package com.github.belyakovleonid.feature_statistics.domain.model
 data class StatisticsCategory(
     val name: String,
     val iconUrl: String,
-    val percent: Double,
+    val percent: Float,
     val color: String,
     val subcategories: List<StatisticsSubcategory>
 )
