+++ feature-statistics/src/main/java/com/github/belyakovleonid/feature_statistics/domain/model/StatisticsSubcategory.kt
@@ -2,5 +2,5 @@ package com.github.belyakovleonid.feature_statistics.domain.model
 
 data class StatisticsSubcategory(
     val name: String,
-    val percent: Double
+    val percent: Float
 )
